task/api/v1/serializers.py
from rest_framework import serializers
from task.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['title','is_done','author','datetime_created']