requirements.txt
# general modules
django >3.2,<3.3
python-decouple


# 3rd party modules
django_extensions

# rest framework
djangorestframework
markdown
django-filter